Documentation and restructuring
