+++ pylana/api.py
@@ -8,27 +8,42 @@ import warnings
 import requests
 
 from .structures import User
-from .user_management import get_user_information
-from .utils import _create_authorization_header
-from .decorators import handle_response
+from .decorators import handle_response, expect_json
 
 
-def handle_http_error(func):
-    @functools.wraps(func)
-    def wrapper(*args, **kwargs):
-        resp = func(*args, **kwargs)
-        try:
-            resp.raise_for_status()
-        except requests.exceptions.HTTPError as e:
-            pass
-            warnings.warn(str(e), Warning)
+def _create_authorization_header(token: str) -> dict:
+    return {"Authorization": f"API-Key {token}"}
 
-        return resp
-    return wrapper
+
+@expect_json
+@handle_response
+def get_user_information(scheme: str, host: str, token: str, port=None) -> dict:
+
+    base_url = f'{scheme}://{host}' + (f':{port}' if port else '')
+    headers = _create_authorization_header(token)
+    r = requests.get(base_url + '/api/users/by-token', headers=headers)
+    r.raise_for_status()
+    return r
+
+
+def get_user(scheme: str, host: str, token: str, port=None) -> User:
+    user_info = get_user_information(scheme, host, token, port)
+    return User(user_id=user_info.get('id'),
+                organization_id=user_info.get('organizationId'),
+                api_key=user_info.get('apiKey'),
+                role=user_info.get('role'))
 
 
 # TODO: consider certificate passing for TLS
 class API:
+    """
+    an api instanced for a specific user and Lana API
+
+    Attributes
+        url (str): the base url of the api (scheme, host and port)
+        user (User): a User dataclass encapsulating the user of the api information
+        headers (dict): the authorization header
+    """
 
     def __init__(self, scheme: str, host: str, token: str, port: int = None):
 
@@ -39,9 +54,6 @@ class API:
                          api_key=user_info.get('apiKey'),
                          role=user_info.get('role'))
         self.headers = _create_authorization_header(token)
-
-        # legacy
-        self.user_info = user_info
 
     def _request(self, method, route, headers=None, additional_headers=None, **kwargs):
         headers = {**self.headers, **(additional_headers or dict()), **(headers or dict())}
