+++ pylana/logs.py
@@ -195,7 +195,7 @@ class LogsAPI(ResourceAPI):
         """
 
         log_id = log_id or self.get_log_id(log_name)
-        resp = self.get_event_csv(log_id)
+        resp = self.request_event_csv(log_id)
         csv_stream = io.StringIO(resp.text)
         return pd.read_csv(csv_stream, dtype='object', **kwargs)
